# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hohaya API",
        "description": "API documentation for Hohaya platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Mon profil",
                "operationId": "695a5ba8ff7a490aad2bed007b5499f9",
                "responses": {
                    "200": {
                        "description": "Profil utilisateur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Mettre à jour mon profil",
                "operationId": "ef8cff62e38652f6c244cdc2f0cdadb4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "adress": {
                                        "type": "string"
                                    },
                                    "preferences": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/avatar": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Upload avatar",
                "operationId": "3a24ea04ae96264eee7e82e1e93e936d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar mis à jour"
                    },
                    "422": {
                        "description": "Fichier invalide"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Supprimer avatar",
                "operationId": "1273a089f49cf742902ca57976a630e6",
                "responses": {
                    "200": {
                        "description": "Avatar supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/account": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Supprimer mon compte",
                "operationId": "7b3e247012567b8ba95773a7716c4811",
                "responses": {
                    "200": {
                        "description": "Compte supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/restore/{id}": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Restaurer un compte supprimé (admin)",
                "operationId": "ad22bc167f4e55cc90fd5e02ce2e4975",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte restauré"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Compte introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}